Moved to com.mongodb.client.gridfs namespace
