+++ driver-core/src/main/com/mongodb/client/gridfs/model/GridFSDownloadByNameOptions.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.mongodb.client.model;
+package com.mongodb.client.gridfs.model;
 
 /**
  * The GridFS download by name options
