+++ driver/src/main/com/mongodb/client/gridfs/GridFSDownloadStream.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.mongodb.client;
+package com.mongodb.client.gridfs;
 
 import org.bson.BsonDocument;
