+++ driver/src/main/com/mongodb/client/gridfs/GridFSDownloadStreamImpl.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.mongodb;
+package com.mongodb.client.gridfs;
 
-import com.mongodb.client.GridFSDownloadStream;
+import com.mongodb.MongoGridFSException;
 import com.mongodb.client.MongoCollection;
 import org.bson.BsonDocument;
 import org.bson.BsonInt32;
