+++ driver/src/main/com/mongodb/client/gridfs/GridFSUploadStreamImpl.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.mongodb;
+package com.mongodb.client.gridfs;
 
-import com.mongodb.client.GridFSUploadStream;
+import com.mongodb.MongoGridFSException;
 import com.mongodb.client.MongoCollection;
 import org.bson.BsonBinary;
 import org.bson.BsonDateTime;
